+++ database/new_types.py
@@ -1,8 +1,34 @@
-from collections import namedtuple
-from dataclasses import make_dataclass
+from dataclasses import dataclass
+import datetime as dt
 
 # see create_tables.sql for type info
-User = make_dataclass('User', ['id', 'name'])
-Event = make_dataclass('Event', ['token', 'name'])
-Expense = make_dataclass('Expense', ['id', 'name', 'sum', 'lender_id', 'event_token', 'datetime'])
-Debt = make_dataclass('Debt', ['expense_id', 'lender_id', 'debtor_id', 'sum'])
+
+
+@dataclass
+class User:
+    id:   int = None
+    name: str = None
+
+
+@dataclass
+class Event:
+    token: str = None
+    name:  str = None
+
+
+@dataclass
+class Expense:
+    id:          int = None
+    name:        str = None
+    sum:         int = None
+    lender_id:   int = None
+    event_token: str = None
+    datetime:    dt.datetime = None
+
+
+@dataclass
+class Debt:
+    expense_id: int = None
+    lender_id:  int = None
+    debtor_id:  int = None
+    sum:        int = None
